Add files via upload
1. Project Overview and Objectives
The main purpose of this project was to build a CNN model that would classify if subject has a tumor or not base on MRI scan.

1.1. Data Set Description
The image data that was used for this problem is Brain MRI Images for Brain Tumor Detection. It conists of MRI scans of two classes:

NO - no tumor, encoded as 0
YES - tumor, encoded as 1
Unfortunately, the data set description doesn't hold any information where this MRI scans come from and so on.

1.2. What is Brain Tumor?
A brain tumor occurs when abnormal cells form within the brain. There are two main types of tumors: cancerous (malignant) tumors and benign tumors. Cancerous tumors can be divided into primary tumors, which start within the brain, and secondary tumors, which have spread from elsewhere, known as brain metastasis tumors. All types of brain tumors may produce symptoms that vary depending on the part of the brain involved. These symptoms may include headaches, seizures, problems with vision, vomiting and mental changes. The headache is classically worse in the morning and goes away with vomiting. Other symptoms may include diffic